New exception method with catches the actual error this time for entering a number rather than a boolean or any other invalid format errors, debug also ,
