--- src/main/java/hello/exceptions/ApiGlobal.java
@@ -1,14 +1,0 @@
-package hello.exceptions;
-
-public class ApiGlobal extends Exception {
-
-    public ApiGlobal(String message) {
-        super(message);
-    }
-
-    public ApiGlobal(Throwable cause) {
-        super(cause);
-    }
-
-
-}
